Exam/2025/sample1/rest/requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv

